+++ src/main/java/chess/piece/Color.java
@@ -1,4 +1,4 @@
-package chess;
+package chess.piece;
 
 public enum Color {
     BLACK, WHITE
