+++ src/main/java/chess/piece/Piece.java
@@ -1,4 +1,6 @@
-package chess;
+package chess.piece;
+
+import chess.position.Position;
 
 public abstract class Piece {
