+++ src/test/java/chess/piece/TestPiece.java
@@ -1,5 +1,7 @@
-package chess;
+package chess.piece;
 
+
+import chess.position.Position;
 
 public class TestPiece extends Piece {
