+++ src/test/java/chess/position/RankTest.java
@@ -1,4 +1,4 @@
-package chess;
+package chess.position;
 
 import static org.assertj.core.api.Assertions.assertThat;
